CommunityTests: split into separate projects
The tests run too long, so let's split them apart so we can run them in
parallel.
