+++ scalafmt-tests-community/common/src/main/scala/org/scalafmt/community/common/CommunityBuild.scala
@@ -1,4 +1,4 @@
-package org.scalafmt.community
+package org.scalafmt.community.common
 
 import scala.meta._
