+++ scalafmt-tests-community/common/src/main/scala/org/scalafmt/community/common/TestStats.scala
@@ -1,4 +1,4 @@
-package org.scalafmt.community
+package org.scalafmt.community.common
 
 case class TestStats(
     checkedFiles: Int,
