+++ scalafmt-tests-community/other/src/test/scala/org/scalafmt/community/other/CommunityMunitSuite.scala
@@ -1,4 +1,6 @@
-package org.scalafmt.community
+package org.scalafmt.community.other
+
+import org.scalafmt.community.common.CommunityRepoSuite
 
 import scala.meta._
