+++ scalafmt-tests-community/scala2/src/test/scala/org/scalafmt/community/scala2/CommunityScala2Suite.scala
@@ -1,4 +1,6 @@
-package org.scalafmt.community
+package org.scalafmt.community.scala2
+
+import org.scalafmt.community.common.CommunityRepoSuite
 
 import scala.meta._
