+++ scalafmt-tests-community/scala3/src/test/scala/org/scalafmt/community/scala3/CommunityScala3Suite.scala
@@ -1,4 +1,6 @@
-package org.scalafmt.community
+package org.scalafmt.community.scala3
+
+import org.scalafmt.community.common.CommunityRepoSuite
 
 import scala.meta._
